feat: 전체 엔티티 JPA 전환 및 DB 스키마 개선 #7
## 변경사항
- 11개 도메인 엔티티 JPA 애노테이션 추가
- BaseTimeEntity 추상 클래스 생성 (User, Product, Order, Coupon, CartItem, ProductStatistics 적용)
- 낙관적 락 적용 (Product, Point, Coupon - @Version)
- TransmissionStatus Enum 추가

## DB 스키마 개선
- cart_items: updated_at 컬럼 추가
- data_transmissions: error_message, last_error_at 컬럼 추가
- 모든 금액 필드 INT → BIGINT 통일

## 타입 및 참조 개선
- Point 관련 모든 금액 필드 Integer → Long 변경
- PointTransaction: userId → pointId 참조로 변경 (Point 테이블 직접 참조)

+++ src/main/java/com/hhplus/hhplus_ecommerce/order/application/OrderService.java
@@ -63,7 +63,7 @@ public class OrderService {
             orderItems.add(orderItem);
         }
 
-        Integer discountAmount = 0;
+        Long discountAmount = 0L;
         if (couponId != null) {
             UserCoupon userCoupon = userCouponRepository.findByUserIdAndCouponId(userId, couponId)
                     .orElseThrow(() -> new BusinessException(ErrorCode.COUPON_NOT_FOUND));
@@ -73,8 +73,8 @@ public class OrderService {
             }
 
             // 주문 총액 계산
-            int totalAmount = orderItems.stream()
-                    .mapToInt(OrderItem::getSubtotal)
+            long totalAmount = orderItems.stream()
+                    .mapToLong(OrderItem::getSubtotal)
                     .sum();
             discountAmount = userCoupon.calculateDiscount(totalAmount);
         }
